# backend/docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "娱乐中心系统API",
    "description": "\n        ## 娱乐中心系统API文档\n        \n        这是一个基于FastAPI的娱乐中心系统，提供以下功能：\n        \n        ### 🎮 游戏模块\n        - **刮刮乐游戏**: 多种主题的刮刮乐游戏\n        - **老虎机游戏**: 经典和现代老虎机\n        - **幸运大转盘**: 多种类型的转盘游戏\n        \n        ### 👤 用户管理\n        - 用户注册和登录\n        - 积分管理\n        - 游戏历史记录\n        \n        ### 📊 数据统计\n        - 用户游戏统计\n        - 排行榜系统\n        - 实时游戏状态\n        \n        ### 🔧 管理后台\n        - 用户管理\n        - 游戏配置\n        - 数据分析\n        \n        ### 🔐 认证方式\n        使用JWT Bearer Token进行认证。在请求头中添加：\n        ```\n        Authorization: Bearer <your_token>\n        ```\n        \n        ### 📱 响应格式\n        所有API响应都采用JSON格式，错误响应包含以下字段：\n        - `error`: 是否为错误\n        - `message`: 错误消息\n        - `status_code`: HTTP状态码\n        ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/games/scratch-card/play": {
      "post": {
        "tags": [
          "游戏"
        ],
        "summary": "Play Scratch Card",
        "description": "玩刮刮乐游戏",
        "operationId": "play_scratch_card_api_games_scratch_card_play_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScratchCardPlayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScratchCardPlayResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/games/scratch-card/scratch": {
      "post": {
        "tags": [
          "游戏"
        ],
        "summary": "Scratch Area",
        "description": "刮开指定区域",
        "operationId": "scratch_area_api_games_scratch_card_scratch_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "area_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Area Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "Card Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Scratch Area Api Games Scratch Card Scratch Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/games/slot-machine/play": {
      "post": {
        "tags": [
          "游戏"
        ],
        "summary": "Play Slot Machine",
        "description": "玩老虎机游戏",
        "operationId": "play_slot_machine_api_games_slot_machine_play_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SlotMachinePlayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlotMachinePlayResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/games/wheel-fortune/play": {
      "post": {
        "tags": [
          "游戏"
        ],
        "summary": "Play Wheel Fortune",
        "description": "玩幸运大转盘游戏",
        "operationId": "play_wheel_fortune_api_games_wheel_fortune_play_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WheelFortunePlayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WheelFortunePlayResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/games/history": {
      "get": {
        "tags": [
          "游戏"
        ],
        "summary": "Get Game History",
        "description": "获取游戏历史记录",
        "operationId": "get_game_history_api_games_history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "game_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Game Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GameHistoryResponse"
                  },
                  "title": "Response Get Game History Api Games History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GameHistoryResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "game_type": {
            "type": "string",
            "title": "Game Type"
          },
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "bet_amount": {
            "type": "integer",
            "title": "Bet Amount"
          },
          "win_amount": {
            "type": "integer",
            "title": "Win Amount"
          },
          "net_win": {
            "type": "integer",
            "title": "Net Win"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "result_data": {
            "type": "object",
            "title": "Result Data"
          }
        },
        "type": "object",
        "required": [
          "id",
          "game_type",
          "template_id",
          "bet_amount",
          "win_amount",
          "net_win",
          "created_at",
          "result_data"
        ],
        "title": "GameHistoryResponse",
        "description": "游戏历史记录响应"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScratchCardPlayRequest": {
        "properties": {
          "template_id": {
            "type": "string",
            "title": "Template Id",
            "description": "模板ID"
          }
        },
        "type": "object",
        "required": [
          "template_id"
        ],
        "title": "ScratchCardPlayRequest",
        "description": "刮刮乐游戏请求"
      },
      "ScratchCardPlayResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "card_data": {
            "type": "object",
            "title": "Card Data"
          },
          "user_credits": {
            "type": "integer",
            "title": "User Credits"
          },
          "game_record_id": {
            "type": "integer",
            "title": "Game Record Id"
          }
        },
        "type": "object",
        "required": [
          "success",
          "card_data",
          "user_credits",
          "game_record_id"
        ],
        "title": "ScratchCardPlayResponse",
        "description": "刮刮乐游戏响应"
      },
      "SlotMachinePlayRequest": {
        "properties": {
          "template_id": {
            "type": "string",
            "title": "Template Id",
            "description": "模板ID"
          },
          "bet_lines": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bet Lines",
            "description": "下注线数，不指定则使用全部支付线"
          }
        },
        "type": "object",
        "required": [
          "template_id"
        ],
        "title": "SlotMachinePlayRequest",
        "description": "老虎机游戏请求"
      },
      "SlotMachinePlayResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "result": {
            "type": "object",
            "title": "Result"
          },
          "user_credits": {
            "type": "integer",
            "title": "User Credits"
          },
          "game_record_id": {
            "type": "integer",
            "title": "Game Record Id"
          }
        },
        "type": "object",
        "required": [
          "success",
          "result",
          "user_credits",
          "game_record_id"
        ],
        "title": "SlotMachinePlayResponse",
        "description": "老虎机游戏响应"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WheelFortunePlayRequest": {
        "properties": {
          "template_id": {
            "type": "string",
            "title": "Template Id",
            "description": "模板ID"
          }
        },
        "type": "object",
        "required": [
          "template_id"
        ],
        "title": "WheelFortunePlayRequest",
        "description": "幸运大转盘游戏请求"
      },
      "WheelFortunePlayResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "result": {
            "type": "object",
            "title": "Result"
          },
          "user_credits": {
            "type": "integer",
            "title": "User Credits"
          },
          "game_record_id": {
            "type": "integer",
            "title": "Game Record Id"
          }
        },
        "type": "object",
        "required": [
          "success",
          "result",
          "user_credits",
          "game_record_id"
        ],
        "title": "WheelFortunePlayResponse",
        "description": "幸运大转盘游戏响应"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "开发服务器"
    },
    {
      "url": "https://api.entertainment-center.com",
      "description": "生产服务器"
    }
  ],
...
}